restructured all ccc files under corporate-creditcard folder with all imports
