+++ common/functions.js
@@ -23,7 +23,7 @@ import {
   documentUpload,
   checkMode,
   customSetFocus,
-} from '../creditcards/corporate-creditcardFunctions.js';
+} from '../creditcards/corporate-creditcard/corporate-creditcardFunctions.js';
 
 import {
   validatePan,
@@ -41,7 +41,7 @@ import {
 } from './makeRestAPI.js';
 
 import * as CONSTANT from './constants.js';
-import * as CC_CONSTANT from '../creditcards/constant.js';
+import * as CC_CONSTANT from '../creditcards/corporate-creditcard/constant.js';
 
 const { ENDPOINTS } = CONSTANT;
 const { JOURNEY_NAME } = CC_CONSTANT;
