+++ common/journey-utils.js
@@ -8,7 +8,7 @@ import {
 import { fetchJsonResponse } from './makeRestAPI.js';
 
 import * as CONSTANT from './constants.js';
-import * as CC_CONSTANT from '../creditcards/constant.js';
+import * as CC_CONSTANT from '../creditcards/corporate-creditcard/constant.js';
 
 const { ENDPOINTS, CHANNEL } = CONSTANT;
 const { JOURNEY_NAME } = CC_CONSTANT;
