+++ creditcards/corporate-creditcard/cc.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-tabs */
 /* eslint no-console: ["error", { allow: ["warn", "error", "debug"] }] */
-import openModal from '../blocks/modal/modal.js';
+import openModal from '../../blocks/modal/modal.js';
 
 const createLabelInElement = (elementSelector, labelClass) => {
   /**
