+++ creditcards/corporate-creditcard/creditcards.css
@@ -1,4 +1,4 @@
-@import url('../styles/color-palette.css');
+@import url('../../styles/color-palette.css');
 
 main {
   form {
